requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
transformers==4.38.2
torch==2.2.0
pandas==2.1.4
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
tqdm==4.66.1
python-dotenv==1.0.0